+++ build.rs
@@ -21,5 +21,6 @@ fn main() {
         println!("cargo:rustc-cfg=feature=\"rstr_task\"");
         println!("cargo:rustc-cfg=feature=\"messagebuf\"");
         println!("cargo:rustc-cfg=feature=\"ovrhdr\"");
+        println!("cargo:rustc-cfg=feature=\"subprio\"");
     }
 }