+++ src/task.rs
@@ -10,6 +10,7 @@ use crate::{
 // TODO: sta_ovr
 // TODO: stp_ovr
 // TODO: ref_ovr
+// TODO: chg_spr
 
 define_error_kind! {
     /// Error type for [`TaskRef::activate`].