+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **Added:** `itron::semaphore::Semaphore[Ref]`
 - **Added:** `itron::abi::{*_mbf}` (message buffers) and Cargo feature `messagebuf`
 - **Added:** `itron::abi::{*_ovr}` (overrun handlers) and Cargo feature `ovrhdr`
+- **Added:** `itron::abi::chg_spr` (change task subpriority) and Cargo feature `subprio`
 
 ## 0.1.0 - 2021-06-15